ARM: omap4: tuna: fix JIG attach/detach handling
If a JIG cable is attached and then detached an OTG_EVENT_NONE
event will be sent to the USB driver on the detach.  This creates
an inbalance in the USB driver's PM usage count.  This change
tracks when a JIG cable is attached so that the event will not be
generated on detach.

Change-Id: I6f0e992f557299662201f98e802d65e1780ab01e
Signed-off-by: Adam Hampson <ahampson@sta.samsung.com>