OMAP4: tuna: gps: keep rts line under hw flow control
GPS seems to be using sw flow control initially,
manually keeping the RTS line high.

After that, the next sequence seems to be to use HW flow control.
In this state, RTS line should behave as per flow control spec.
This will not happen, if re-mux is not set in pad configuration

Change-Id: Ie2a0edca6acef6b4baa7f471fd09c8ef428db437
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Mike Lockwood <lockwood@android.com>