input: mms_ts: disable irqs during time sensitive operations
While flashing the new firmware, the touch IC is booted into a
separate mode and the firmware is bit banged over the
SCL/SDA/nINT(resetb) pins. Some operations are very timing
sensitive, and we just cannot guarantee that we can meet them.

It looks like, with the current delays, we can keep irqs off
for around 150uS-200uS.

This is unfortunate, but it is only done during firmware reflash
on boot.

Change-Id: If74de1931f5fb0ed0f8e5e2e1e636d515e1e77a5
Signed-off-by: Dima Zavin <dima@android.com>